Rename `telemetry_reconciler` files to `cassandra_telemetry_reconciler` (we will place stargate telemetry under the stargate controller). Factor out `NewFakeClientWRestMapper()` related testing helpers into `pkg/test` so they can be shared between Cassandra + Stargate telemetry. Move `NewFakeClient()` out from `testObjects.go` into a new client-focused testing helper file. Rename `testObjects.go` `test_objects.go` to accord with Go naming conventions.
