+++ controllers/k8ssandra/cassandra_telemetry_reconciler.go
@@ -5,7 +5,6 @@ package k8ssandra
 import (
 	"context"
 	"errors"
-
 	"github.com/go-logr/logr"
 	cassdcapi "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	api "github.com/k8ssandra/k8ssandra-operator/apis/k8ssandra/v1alpha1"
