+++ pkg/test/test_objects.go
@@ -4,16 +4,10 @@ package test
 import (
 	cassdcapi "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	k8ssandraapi "github.com/k8ssandra/k8ssandra-operator/apis/k8ssandra/v1alpha1"
-	reaperapi "github.com/k8ssandra/k8ssandra-operator/apis/reaper/v1alpha1"
 	stargateapi "github.com/k8ssandra/k8ssandra-operator/apis/stargate/v1alpha1"
-	promapi "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
 
 // NewK8ssandraCluster returns a minimum viable k8ssandra cluster.
@@ -50,24 +44,6 @@ func NewK8ssandraCluster(name string, namespace string) k8ssandraapi.K8ssandraCl
 
 }
 
-// NewFakeClient gets a fake client loaded up with a scheme that contains all the APIs used in this project.
-func NewFakeClient() (client.Client, error) {
-	schemeBuilder := scheme.Builder{}
-	testScheme, err := schemeBuilder.Build()
-	if err != nil {
-		return nil, err
-	}
-	utilruntime.Must(promapi.AddToScheme(testScheme))
-	utilruntime.Must(cassdcapi.AddToScheme(testScheme))
-	utilruntime.Must(k8ssandraapi.AddToScheme(testScheme))
-	utilruntime.Must(reaperapi.AddToScheme(testScheme))
-	utilruntime.Must(stargateapi.AddToScheme(testScheme))
-	fakeClient := fake.NewClientBuilder().
-		WithScheme(testScheme).
-		Build()
-	return fakeClient, nil
-}
-
 func NewCassandraDatacenter() cassdcapi.CassandraDatacenter {
 	return cassdcapi.CassandraDatacenter{
 		TypeMeta: metav1.TypeMeta{
@@ -94,6 +70,29 @@ func NewCassandraDatacenter() cassdcapi.CassandraDatacenter {
 				},
 			},
 			ClusterName: "test-cluster",
+		},
+	}
+}
+
+func NewStargate(name string, namespace string) stargateapi.Stargate {
+	return stargateapi.Stargate{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "stargate.k8ssandra.io/v1alpha1",
+			Kind:       "Stargate",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: stargateapi.StargateSpec{
+			StargateDatacenterTemplate: stargateapi.StargateDatacenterTemplate{
+				StargateClusterTemplate: stargateapi.StargateClusterTemplate{
+					StargateTemplate: stargateapi.StargateTemplate{
+						AllowStargateOnDataNodes: true,
+					},
+					Size: 1,
+				},
+			},
 		},
 	}
 }
